requirements.txt
llama-index-core>=0.12.0
llama-index-llms-anthropic>=0.6.0
llama-index-embeddings-openai>=0.3.0
llama-parse>=0.5.0
pinecone>=5.0.0
pinecone-text>=0.9.0
anthropic>=0.40.0
openai>=1.50.0
streamlit>=1.38.0
sentence-transformers>=3.0.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
